refactor dataclasses, parquet processing
