+++ src/nwb_benchmarks/database/__init__.py
@@ -1,0 +1,22 @@
+"""Exposed imports to the `database` submodule."""
+
+from ._models import (
+    Machine,
+    Result,
+    Results,
+    Environment
+)
+
+from ._processing import (
+    concat_dataclasses_to_parquet,
+    repackage_as_parquet,
+)
+
+__all__ = [
+    "Machine",
+    "Result",
+    "Results",
+    "Environment",
+    "concat_dataclasses_to_parquet",
+    "repackage_as_parquet",
+]
